… README.md
… sloth/legacy/writer.py
… sloth/parser.py
… sloth/wrappers.py
… sloth/writer.py
… tests/test_main.py